app/requirements.txt
console-menu==0.8.0
web3==5.31.3
py-solc-x==1.1.1
python-dotenv==0.21.1
prettytable==3.6.0